Add mastering IPython 4 on the book section of the website.
As requested on steering council mailing list.

Also add a "Book" section to the navbar as I took me a couple of minutes
to figure out how to get there without changing th url directly.

I'd like to link to the author profile, but I'm unsure which url to use.

The description is a subset of the description on Packt website.

This automatically add the book in the rotating book in the sidebar.