Add recent packt book and videos
I got several requests, and while I would prefer for them to send the
pull-request directly I can't expect marketing people to do that.

I would like for packt to not be the only publisher on this page as
well.

Also tweak a bit the CSS for the page to look nicer.