Update index.rst
This small change corrects two links on the main page.

The first is a link to the official examples, which currently point to an outdated location.

The second is a link to the notebook gallery. The old link worked, but pointed to a github page which indicated that the resource had moved and pointed to a second page, which **also** indicated that the resource had moved and pointed to yet another page. I've replaced the old link with a direct link to the current location.

I note that these were brought up at https://github.com/ipython/ipython/issues/13766, and I'm fixing them here.